from mac

+++ lib/module_authorization/service/auth_service.dart
@@ -226,10 +226,13 @@ Future<AppUser>  getCurrentUser()async{
   }
 
   void fakeAccount() async{
-  print('fake account');
- String uid = await FirebaseAuth.instance.currentUser!.uid;
- await _repository.deleteFakeProfile(uid);
+  if(FirebaseAuth.instance.currentUser != null){
+    String uid = await FirebaseAuth.instance.currentUser!.uid;
+    await _repository.deleteFakeProfile(uid);
     FirebaseAuth.instance.currentUser!.delete();
+  }
+
+
   }
 
  Future<AuthResponse> resetPassword(String email) async{
